tukqph20340_admin_gd1

+++ src/main/java/com/example/demo/repository/NhaXuatbanRepository.java
@@ -4,5 +4,5 @@ import com.example.demo.model.NhaXuatBan;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface NhaXuatbanRepository extends JpaRepository<NhaXuatBan,Integer> {
+public interface NhaXuatbanRepository extends JpaRepository<NhaXuatBan,Long> {
 }
