+++ src/main/java/com/example/demo/repository/SachRepository.java
@@ -4,5 +4,5 @@ import com.example.demo.model.Sach;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface SachRepository extends JpaRepository<Sach,Integer> {
+public interface SachRepository extends JpaRepository<Sach,Long> {
 }
