+++ src/main/java/com/example/demo/repository/TacGiaRepository.java
@@ -4,5 +4,5 @@ import com.example.demo.model.TacGia;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface TacGiaRepository extends JpaRepository<TacGia,Integer> {
+public interface TacGiaRepository extends JpaRepository<TacGia,Long> {
 }
