+++ src/main/java/com/example/demo/service/NhaXuatBanSevice.java
@@ -7,8 +7,8 @@ import org.springframework.data.domain.Pageable;
 import java.util.ArrayList;
 public interface NhaXuatBanSevice {
     ArrayList<NhaXuatBan> getAll();
-    void delete(Integer id);
-    NhaXuatBan detail(Integer id);
+    void delete(Long id);
+    NhaXuatBan detail(Long id);
     NhaXuatBan add(NhaXuatBan nhaXuatBan);
     NhaXuatBan update(NhaXuatBan nhaXuatBan);
     Page<NhaXuatBan> phanTrang(Pageable pageable);
