+++ src/main/java/com/example/demo/service/SachSevice.java
@@ -7,8 +7,8 @@ import org.springframework.data.domain.Pageable;
 import java.util.ArrayList;
 public interface SachSevice {
     ArrayList<Sach> getAll();
-    void delete(Integer id);
-    Sach detail(Integer id);
+    void delete(Long id);
+    Sach detail(Long id);
     Sach add(Sach sach);
     Sach update(Sach sach);
     Page<Sach> phanTrang(Pageable pageable);
